move test email dropdown into its own component, fix e2e tests
