+++ apps/admin-x-settings/src/components/settings/membership/member-emails/welcome-email-modal.tsx
@@ -1,15 +1,13 @@
 import NiceModal from '@ebay/nice-modal-react';
-import validator from 'validator';
 import {useEffect, useRef, useState} from 'react';
 
 import MemberEmailEditor from './member-email-editor';
 import {Button, Hint, Modal, TextField} from '@tryghost/admin-x-design-system';
 import {useForm, useHandleError} from '@tryghost/admin-x-framework/hooks';
 
-import {JSONError} from '@tryghost/admin-x-framework/errors';
+import TestEmailDropdown from './test-email-dropdown';
 import {getSettingValues} from '@tryghost/admin-x-framework/api/settings';
-import {useCurrentUser} from '@tryghost/admin-x-framework/api/current-user';
-import {useEditAutomatedEmail, useSendTestWelcomeEmail} from '@tryghost/admin-x-framework/api/automated-emails';
+import {useEditAutomatedEmail} from '@tryghost/admin-x-framework/api/automated-emails';
 import {useGlobalData} from '../../../../components/providers/global-data-provider';
 import {useRouting} from '@tryghost/admin-x-framework/routing';
 import type {AutomatedEmail} from '@tryghost/admin-x-framework/api/automated-emails';
@@ -46,15 +44,9 @@ const isEmptyLexical = (lexical: string | null | undefined): boolean => {
 
 const WelcomeEmailModal = NiceModal.create<WelcomeEmailModalProps>(({emailType = 'free', automatedEmail}) => {
     const {updateRoute} = useRouting();
-    const {data: currentUser} = useCurrentUser();
     const {mutateAsync: editAutomatedEmail} = useEditAutomatedEmail();
-    const {mutateAsync: sendTestEmail} = useSendTestWelcomeEmail();
     const [showTestDropdown, setShowTestDropdown] = useState(false);
-    const [testEmail, setTestEmail] = useState(currentUser?.email || '');
-    const [testEmailError, setTestEmailError] = useState('');
-    const [sendState, setSendState] = useState<'idle' | 'sending' | 'sent'>('idle');
     const dropdownRef = useRef<HTMLDivElement>(null);
-    const sendStateTimeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null);
     const handleError = useHandleError();
     const {settings} = useGlobalData();
     const [siteTitle, defaultEmailAddress] = getSettingValues<string>(settings, ['title', 'default_email_address']);
@@ -83,13 +75,6 @@ const WelcomeEmailModal = NiceModal.create<WelcomeEmailModalProps>(({emailType =
             return newErrors;
         }
     });
-
-    // Update test email when current user data loads
-    useEffect(() => {
-        if (currentUser?.email) {
-            setTestEmail(currentUser.email);
-        }
-    }, [currentUser?.email]);
 
     // Close dropdown when clicking outside
     useEffect(() => {
@@ -126,46 +111,6 @@ const WelcomeEmailModal = NiceModal.create<WelcomeEmailModalProps>(({emailType =
         };
     }, []);
 
-    useEffect(() => () => clearTimeout(sendStateTimeoutRef.current ?? undefined), []);
-
-    const handleSendTestEmail = async () => {
-        setTestEmailError('');
-
-        if (!validator.isEmail(testEmail)) {
-            setTestEmailError('Please enter a valid email address');
-            return;
-        }
-
-        // check that subject and lexical are valid
-        if (!validate()) {
-            setTestEmailError('Please complete the required fields');
-            return;
-        }
-
-        setSendState('sending');
-
-        try {
-            await sendTestEmail({
-                id: automatedEmail.id,
-                email: testEmail,
-                subject: formState.subject,
-                lexical: formState.lexical
-            });
-            setSendState('sent');
-            clearTimeout(sendStateTimeoutRef.current!);
-            sendStateTimeoutRef.current = setTimeout(() => setSendState('idle'), 2000);
-        } catch (error) {
-            setSendState('idle');
-            let message;
-            if (error instanceof JSONError && error.data && error.data.errors[0]) {
-                message = error.data.errors[0].context || error.data.errors[0].message;
-            } else if (error instanceof Error) {
-                message = error.message;
-            }
-            setTestEmailError(message || 'Failed to send test email');
-        }
-    };
-
     const senderEmail = automatedEmail?.sender_email || defaultEmailAddress;
     const replyToEmail = automatedEmail?.sender_reply_to || defaultEmailAddress;
 
@@ -193,27 +138,7 @@ const WelcomeEmailModal = NiceModal.create<WelcomeEmailModalProps>(({emailType =
                                     onClick={() => setShowTestDropdown(!showTestDropdown)}
                                 />
                                 {showTestDropdown && (
-                                    <div className='absolute right-0 top-full z-10 mt-2 w-[260px] rounded border border-grey-200 bg-white p-4 shadow-lg'>
-                                        <div className='mb-3'>
-                                            <label className='mb-2 block text-sm font-semibold'>Send test email</label>
-                                            <TextField
-                                                className='!h-[36px]'
-                                                placeholder='you@yoursite.com'
-                                                value={testEmail}
-                                                onChange={(e) => {
-                                                    setTestEmail(e.target.value);
-                                                }}
-                                            />
-                                        </div>
-                                        <Button
-                                            className='w-full'
-                                            color={sendState === 'sent' ? 'green' : 'black'}
-                                            disabled={sendState === 'sending'}
-                                            label={sendState === 'sent' ? 'Sent' : sendState === 'sending' ? 'Sending...' : 'Send'}
-                                            onClick={handleSendTestEmail}
-                                        />
-                                        {testEmailError && <Hint className='mt-2' color='red'>{testEmailError}</Hint>}
-                                    </div>
+                                    <TestEmailDropdown automatedEmailId={automatedEmail.id} lexical={formState.lexical} subject={formState.subject} validateForm={validate} />
                                 )}
                             </div>
                             <Button
